server/requirements.txt
alembic==1.14.0
attr==0.3.2
ConfigParser==7.1.0
contextlib2==21.6.0
cryptography==43.0.3
docutils==0.21.2
filelock==3.16.1
Flask==3.0.3
Flask_Bcrypt==1.0.1
Flask_Cors==4.0.1
Flask_Migrate==4.0.7
flask_session==0.8.0
flask_sqlalchemy==3.1.1
HTMLParser==0.0.2
importlib_metadata==8.5.0
ipython==8.12.3
ipywidgets==8.1.5
Jinja2==3.1.4
jnius==1.1.0
keyring==25.5.0
mock==5.1.0
Pillow==11.0.0
protobuf==5.28.3
pyOpenSSL==24.2.1
python-dotenv==1.0.1
railroad==0.5.0
redis==5.2.0
Sphinx==8.1.3
SQLAlchemy==2.0.29
sqlalchemy_serializer==1.4.22
thread==2.0.5
tornado==6.4.1
urllib3_secure_extra==0.1.0
xmlrpclib==1.0.1
